Finish project "Landing Page"
Use as many concepts as learned to master them.

Create a good structure in HTML to make it
easy to read and easy to edit or create style.

Make CSS as simple as possible.
Use classes and id's to make it easier to code.

Don't really know what to write here, but one
thing is clear: this website is usable. :)